Don't create a separate ENVDIR, use WORKDIR/testenv instead

+++ tests/generate_testenv_pkgs.sh
@@ -27,29 +27,17 @@ echo "=== Created working directory $WORKDIR"
 # Cleanup on exit
 function on_exit() {
 	popd >/dev/null
-	if test "$ENVDIR" != ""; then
-	      echo "=== Removing test environment $ENVDIR"
-	      rm -rf "$ENVDIR"
-	fi
-	if test "$WORKDIR" != ""; then
-	      echo "=== Removing working directory $WORKDIR"
-	      rm -rf "$WORKDIR"
+	if test "$WORKDIR" != "" && test "$KEEP_WORKDIR" == ""; then
+		echo "=== Removing working directory $WORKDIR"
+		rm -rf "$WORKDIR"
 	fi
 }
 trap on_exit EXIT
 
-# Create temporary directory to hold a temporary conda environment used to run the tests
-ENVDIR="$(mktemp -d)"
-if ! test -d "$ENVDIR"; then
-	ENVDIR=""
-	echo "Failed to create temporary directory to hold the test environment ($ENVDIR is not a directory):" >&2
-	exit 1
-fi
-echo "=== Creating and activating test environment in $ENVDIR"
-conda create --no-default-packages --yes -p "$ENVDIR"
-source activate "$ENVDIR" 
-
-echo "=== Configuring channels"
+# Setup conda test environment
+echo "=== Creating and activating test environment in $WORKDIR/testenv"
+conda create --no-default-packages --yes -p "$WORKDIR/testenv"
+source activate "$WORKDIR/testenv"
 while read -r channel || [[ -n "$channel" ]]; do
 	conda config --env --add channels "$channel"
 done < "$TESTS/testenv.channels"
